% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_path.R
\name{create_path}
\alias{create_path}
\title{create_path}
\usage{
create_path(data, smooth = TRUE, method = "ksmooth", ...)
}
\arguments{
\item{data}{sf of points in order, joined with nonspatial flight data representing one aircraft over time.}

\item{smooth}{Boolean, default TRUE. which activates the smoothing algorithm from smoothr.}

\item{method}{Optional, smoothing algorithm to be used, default ksmooth. See smoothr documentation.}

\item{...}{Optional parameters for future use.}
}
\value{
sf of linestrings and interpolated nonspatial data.
}
\description{
Convert point flight data into a chained path.
}
\examples{
\dontrun{create_path(data_sf)}

}
