% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(username, password, icao24, start_seconds_ago = NULL, duration = NULL)
}
\arguments{
\item{username}{Your 'OpenSky Network' username.}

\item{password}{Your 'OpenSky Network' password.}

\item{icao24}{Unique ICAO 24-bit address of the transponder in hex string
representation. All letters need to be lower case.}

\item{start_seconds_ago}{Amount of time (in seconds) before current time user wants to start
collecting state vectors. Does not need to be specified if user only wants current and/or future state vectors.}

\item{duration}{Amount of time (in seconds) for which to collect live data. Does not need to be
specified if user only wants past and/or current state vectors}
}
\value{
An sf object of state vectors for every 5-second snapshot, from the start time to end of duration.
}
\description{
Retrieve state vectors for a specific plane over an optionally specified amount of time
}
\examples{
\dontrun{get_data(username = "your_username", password = "your_password", icao24 = "a8699a",
 start_seconds_ago = 600, duration = 300)}

}
