% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_live_data.R
\name{get_live_data}
\alias{get_live_data}
\title{get_live_data}
\usage{
get_live_data(username, password, duration, icao24 = NULL, ...)
}
\arguments{
\item{username}{Your 'OpenSky Network' username.}

\item{password}{Your 'OpenSky Network' password.}

\item{duration}{Amount of time (in seconds) for which to collect live data.}

\item{icao24}{Optional. Unique icao24 identifier for aircraft}

\item{...}{Optional placeholder for future parameters.}
}
\value{
If icao24 is missing, A non-spatial dataframe of all state vectors retrieved during the collection period.
  if icao24 specified, A geocoded sf is returned instead.
}
\description{
Retrieve all flight vectors every 5 seconds
 starting now up to a specified time into the future.
}
\examples{
\dontrun{collect_live_data(username = "your_username", password = "your_password",
 duration = 30)}

}
