% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_past_data.R
\name{get_past_data}
\alias{get_past_data}
\title{get_past_data}
\usage{
get_past_data(username, password, start_seconds_ago, icao24)
}
\arguments{
\item{username}{Your 'OpenSky Network' username.}

\item{password}{Your 'OpenSky Network' password.}

\item{start_seconds_ago}{Amount of time (in seconds) before current time when user wants to start
collecting state vectors.}

\item{icao24}{Unique ICAO 24-bit address of the transponder in hex string
representation. All letters need to be lower case.}
}
\value{
An sf object of state vectors for every 5-second snapshot, from the start time to current time
}
\description{
Retrieve state vectors for a specific plane if the desired start time is before the current time.
}
\examples{
\dontrun{get_past_data(username = "your_username", password = "your_password",
 start_seconds_ago = 600, icao24 = "3c4b26")}

}
