% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_path.R
\name{lonlat2UTM}
\alias{lonlat2UTM}
\title{lonlat2UTM}
\usage{
lonlat2UTM(lonlat)
}
\arguments{
\item{lonlat}{longitude/latitude coordinates}
}
\value{
UTM code
}
\description{
calculates EPSG code associated with any point on the planet, from Lovelace 6.3
}
