% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{plot_map}
\usage{
plot_map(data, variable, view = TRUE)
}
\arguments{
\item{data}{sf of linestrings and interpolated nonspatial data.}

\item{variable}{String of variable that user wants to respresent by color on the linestrings.}

\item{view}{Boolean (default TRUE) that specifies if the map returned should be an interactive view map
or a stationary plot map, the latter of which conains cartographic elements.}
}
\value{
tmap of aircraft path on top of an OpenTopoMap basemap (view) or polygon world map (plot).
}
\description{
Creates and returns a map object displaying aircraft path and chosen variable information.
}
\examples{
\dontrun{plot_map(data, "velocity")}

}
